+++ src/logger/get-null-logger.ts
@@ -1,0 +1,10 @@
+import { Logger } from '../types';
+
+export default (): Logger => {
+    const noop = () => {};
+
+    return {
+        debug: noop,
+        log: noop,
+    };
+};