+++ src/logger/index.ts
@@ -1,8 +1,0 @@
-import { LogEntry } from './../types';
-
-export { default as consoleLogger } from './console';
-export { default as mochaLogger } from './mocha';
-export { default as junitLogger } from './junit';
-export { compactProjectLogs } from './flatten';
-export const getErrorCount = (projectLogs: LogEntry[]) =>
-    Object.values(projectLogs).reduce((acc, i) => acc + ((i.errors && i.errors.length) || 0), 0);