+++ src/stricter.ts
@@ -1,77 +1,48 @@
 import { getConfig } from './config';
 import { getRuleDefinitions, getRuleApplications, filterFilesToProcess } from './rule';
 import { applyProjectRules, readFilesData } from './processor';
-import {
-    consoleLogger,
-    junitLogger,
-    mochaLogger,
-    compactProjectLogs,
-    getErrorCount,
-} from './logger';
+import { getErrorCount } from './reporter';
 import { listFiles } from './utils';
-import { StricterArguments, Reporter } from './types';
-import debug, { measure } from './debug';
+import { StricterArguments } from './types';
 
 export default ({
-    silent = false,
-    reporter = Reporter.CONSOLE,
-    configPath,
+    options: { configPath },
+    reporter,
+    logger: { debug, log },
 }: StricterArguments): number => {
-    const result = measure('Total', () => {
-        debug({
-            silent,
-            reporter,
-            configPath,
-        });
-
-        if (!silent) {
-            console.log('Stricter: Checking...');
-        }
-
-        const config = measure('Read config', () => getConfig(configPath));
-
-        const fileList = measure('Get file list', () => listFiles(config.root, config.exclude));
-
-        const ruleDefinitions = measure('Get rule definitions', () => getRuleDefinitions(config));
-
-        const ruleApplications = measure('Get rule applications', () =>
-            getRuleApplications(config, ruleDefinitions),
-        );
-
-        const filesToProcess = measure('Get files to process', () =>
-            filterFilesToProcess(config.root, fileList, ruleApplications),
-        );
-
-        const filesData = measure('Read files data', () => readFilesData(filesToProcess));
-
-        const projectResult = measure('Apply rules', () =>
-            applyProjectRules(config.root, filesData, ruleApplications),
-        );
-
-        const logs = measure('Massage logs', () => compactProjectLogs(projectResult));
-
-        measure('Write logs', () => {
-            if (reporter === Reporter.MOCHA) {
-                mochaLogger(logs);
-            } else if (reporter === Reporter.JUNIT) {
-                junitLogger(logs);
-            } else {
-                consoleLogger(logs);
-            }
-        });
-
-        const result = measure('Count errors', () => getErrorCount(logs));
-
-        if (!silent) {
-            if (result === 0) {
-                console.log('Stricter: No errors');
-            } else {
-                console.log(`Stricter: ${result} error${result > 1 ? 's' : ''}`);
-            }
-        }
-
-        return result === 0 ? 0 : 1;
+    debug({
+        reporter,
+        configPath,
     });
+
+    log('Checking...');
+
+    debug('Read config');
+    const config = getConfig(configPath);
+
+    debug('Get file list');
+    const fileList = listFiles(config.root, config.exclude);
+
+    debug('Get rule definitions');
+    const ruleDefinitions = getRuleDefinitions(config);
+
+    debug('Get rule applications');
+    const ruleApplications = getRuleApplications(config, ruleDefinitions);
+
+    debug('Get files to process');
+    const filesToProcess = filterFilesToProcess(config.root, fileList, ruleApplications);
+
+    debug('Read files data');
+    const filesData = readFilesData(filesToProcess);
+
+    debug('Apply rules');
+    const projectResult = applyProjectRules(config.root, filesData, ruleApplications);
+
+    debug('Output report');
+    reporter(projectResult);
+
+    debug('Count errors');
+    const result = getErrorCount(projectResult) === 0 ? 0 : 1;
 
     return result;
 };