+++ CHANGELOG.md
@@ -3,6 +3,10 @@
 ## Master
 
 - Bump dependencies
+- Use OS-specific line-breaks
+- Refactoring
+- Improved JUnit reporter
+- Subtle console reporter changes
 
 ## 0.1.1 - JUnit reporter